Revert "Refactor common code in aws-plugin to aws-common (#5877)"
This reverts commit be0ee9849986db3821142f392eec3c7d8a2568b9.
