--- data-prepper-plugins/aws-common/build.gradle
@@ -1,32 +1,0 @@
-dependencies {
-    implementation project(':data-prepper-api')
-    implementation project(':data-prepper-plugins:aws-plugin-api')
-    implementation 'com.fasterxml.jackson.core:jackson-databind'
-    implementation 'io.micrometer:micrometer-core'
-    implementation 'software.amazon.awssdk:auth'
-    implementation 'software.amazon.awssdk:secretsmanager'
-    implementation 'software.amazon.awssdk:sts'
-    implementation 'software.amazon.awssdk:arns'
-    implementation 'org.hibernate.validator:hibernate-validator:8.0.2.Final'
-    testImplementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
-    testImplementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
-    testImplementation project(':data-prepper-test:test-common')
-}
-
-test {
-    useJUnitPlatform()
-}
-
-jacocoTestCoverageVerification {
-    dependsOn jacocoTestReport
-    violationRules {
-        rule {
-            limit {
-                minimum = 1.0
-            }
-        }
-    }
-}
-
-check.dependsOn jacocoTestCoverageVerification
-
