+++ data-prepper-plugins/aws-plugin/src/main/java/org/opensearch/dataprepper/plugins/aws/AwsPluginConfig.java
@@ -6,7 +6,6 @@
 package org.opensearch.dataprepper.plugins.aws;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.opensearch.dataprepper.aws.common.AwsStsConfiguration;
 
 public class AwsPluginConfig {
