+++ data-prepper-plugins/aws-plugin/src/test/java/org/opensearch/dataprepper/plugins/aws/AwsStsConfigurationTest.java
@@ -3,11 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.aws.common;
+package org.opensearch.dataprepper.plugins.aws;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import software.amazon.awssdk.regions.Region;
@@ -21,15 +20,6 @@ import static org.hamcrest.Matchers.notNullValue;
 public class AwsStsConfigurationTest {
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
-
-    @Test
-    void testStsConfigurationWithNullRegion() throws JsonProcessingException {
-        final String defaultConfigurationAsString = "{\"sts_role_arn\": \"arn:aws:iam::123456789012:role/test-role\"}";
-        final AwsStsConfiguration objectUnderTest = OBJECT_MAPPER.readValue(defaultConfigurationAsString, AwsStsConfiguration.class);
-        assertThat(objectUnderTest, notNullValue());
-        assertThat(objectUnderTest.getAwsStsRoleArn(), equalTo("arn:aws:iam::123456789012:role/test-role"));
-        assertThat(objectUnderTest.getAwsRegion(), equalTo(null));
-    }
 
     @ParameterizedTest
     @MethodSource("getRegions")
