+++ settings.gradle
@@ -165,7 +165,6 @@ include 'data-prepper-plugins:user-agent-processor'
 include 'data-prepper-plugins:in-memory-source-coordination-store'
 include 'data-prepper-plugins:aws-plugin-api'
 include 'data-prepper-plugins:aws-plugin'
-include 'data-prepper-plugins:aws-common'
 include 'data-prepper-plugins:obfuscate-processor'
 include 'data-prepper-plugins:parquet-codecs'
 include 'data-prepper-plugins:aws-sqs-common'
